Pass time period on navigation from Digma metrics
