+++ src/components/Insights/InsightsCatalog/FilterPanel/IssuesFilter/types.ts
@@ -18,5 +18,5 @@ export interface IssuesFilterQuery {
   services?: string[];
   criticalityFilter?: IssueCriticality[];
   criticalityFilterInGlobalScope?: IssueCriticality[];
-  lastDays?: number;
+  lastDays?: number | null;
 }
