+++ src/components/Insights/hooks/useInsightsStats.ts
@@ -3,6 +3,7 @@ import { getFeatureFlagValue } from "../../../featureFlags";
 import { useGetInsightsStatsQuery } from "../../../redux/services/digma";
 import { useConfigSelector } from "../../../store/config/useConfigSelector";
 import { useInsightsSelector } from "../../../store/insights/useInsightsSelector";
+import { isNumber } from "../../../typeGuards/isNumber";
 import { FeatureFlag } from "../../../types";
 
 const REFRESH_INTERVAL = 10 * 1000; // in milliseconds
@@ -51,7 +52,10 @@ export const useInsightsStats = ({
       services:
         filteredServices.length > 0 ? filteredServices.join(",") : undefined,
       displayName: search.length > 0 ? search : undefined,
-      lastDays: isIssuesLastDaysFilterEnabled ? lastDays : undefined
+      lastDays:
+        isIssuesLastDaysFilterEnabled && isNumber(lastDays)
+          ? lastDays
+          : undefined
     },
     {
       skip: !environmentId,
