+++ src/components/Insights/index.tsx
@@ -1,24 +1,16 @@
 import type { KeyboardEvent } from "react";
 import { useCallback, useEffect, useState } from "react";
 import { actions as globalActions } from "../../actions";
-import { usePersistence } from "../../hooks/usePersistence";
 import { usePrevious } from "../../hooks/usePrevious";
 import { useConfigSelector } from "../../store/config/useConfigSelector";
-import {
-  initialState as insightsInitialState,
-  type InsightsData
-} from "../../store/insights/insightsSlice";
+import { type InsightsData } from "../../store/insights/insightsSlice";
 import { useInsightsSelector } from "../../store/insights/useInsightsSelector";
 import { useStore } from "../../store/useStore";
-import { isUndefined } from "../../typeGuards/isUndefined";
-import { type InsightType } from "../../types";
-import { areBackendInfosEqual } from "../../utils/areBackendInfosEqual";
 import { RegistrationDialog } from "../common/RegistrationDialog";
 import type { RegistrationFormValues } from "../common/RegistrationDialog/types";
 import { EmptyState } from "./EmptyState";
 import { useInsightsData } from "./hooks/useInsightsData";
 import { InsightsCatalog } from "./InsightsCatalog";
-import type { IssuesFilterQuery } from "./InsightsCatalog/FilterPanel/IssuesFilter/types";
 import { InsightTicketRenderer } from "./InsightTicketRenderer";
 import * as s from "./styles";
 import type {
@@ -27,22 +19,11 @@ import type {
   InsightsProps
 } from "./types";
 
-export const ISSUES_FILTERS_PERSISTENCE_KEY = "issuesFilters";
-
 export const Insights = ({ insightViewType }: InsightsProps) => {
-  const [areFiltersRehydrated, setAreFiltersRehydrated] = useState(false);
-  const [persistedFilters, setPersistedFilters] =
-    usePersistence<IssuesFilterQuery>(
-      ISSUES_FILTERS_PERSISTENCE_KEY,
-      "project"
-    );
-  const previousPersistedFilters = usePrevious(persistedFilters);
-  const { data, isLoading, refresh } = useInsightsData({
-    areFiltersRehydrated
-  });
+  const { data, isLoading, refresh } = useInsightsData();
   const [infoToOpenJiraTicket, setInfoToOpenJiraTicket] =
     useState<InsightTicketInfo<GenericCodeObjectInsight>>();
-  const { backendInfo, userRegistrationEmail, environments, scope } =
+  const { backendInfo, userRegistrationEmail, environments } =
     useConfigSelector();
   const previousUserRegistrationEmail = usePrevious(userRegistrationEmail);
   const [isRegistrationInProgress, setIsRegistrationInProgress] =
@@ -50,41 +31,8 @@ export const Insights = ({ insightViewType }: InsightsProps) => {
   const isRegistrationEnabled = false;
   const isRegistrationRequired =
     isRegistrationEnabled && !userRegistrationEmail;
-  const {
-    setInsightViewType,
-    setInsightsFilteredInsightTypes: setFilteredInsightTypes,
-    setInsightsFilteredInsightTypesInGlobalScope:
-      setFilteredInsightTypesInGlobalScope,
-    setInsightsFilteredCriticalityLevels: setFilteredCriticalityLevels,
-    setInsightsFilteredCriticalityLevelsInGlobalScope:
-      setFilteredCriticalityLevelsInGlobalScope,
-    setInsightsFilters: setFilters,
-    setInsightsLastDays: setLastDays,
-    resetInsights: reset
-  } = useStore.getState();
-  const {
-    insightViewType: storedInsightViewType,
-    filteredInsightTypes: filteredInsightTypesInSpanScope,
-    filteredInsightTypesInGlobalScope,
-    filteredCriticalityLevels: filteredCriticalityLevelsInSpanScope,
-    filteredCriticalityLevelsInGlobalScope,
-    filters,
-    lastDays
-  } = useInsightsSelector();
-  const scopeSpanCodeObjectId = scope?.span?.spanCodeObjectId;
-  const filteredInsightTypes = scopeSpanCodeObjectId
-    ? filteredInsightTypesInSpanScope
-    : filteredInsightTypesInGlobalScope;
-  const previousFilteredInsightTypes = usePrevious(filteredInsightTypes);
-  const filteredCriticalityLevels = scopeSpanCodeObjectId
-    ? filteredCriticalityLevelsInSpanScope
-    : filteredCriticalityLevelsInGlobalScope;
-  const previousFilteredCriticalityLevels = usePrevious(
-    filteredCriticalityLevels
-  );
-  const previousFilters = usePrevious(filters);
-  const previousLastDays = usePrevious(lastDays);
-  const previousBackendInfo = usePrevious(backendInfo);
+  const { setInsightViewType, resetInsights: reset } = useStore.getState();
+  const { insightViewType: storedInsightViewType } = useInsightsSelector();
 
   useEffect(() => {
     return () => {
@@ -107,107 +55,6 @@ export const Insights = ({ insightViewType }: InsightsProps) => {
     userRegistrationEmail,
     isRegistrationInProgress,
     previousUserRegistrationEmail
-  ]);
-
-  // Rehydrate filters from persistence
-  useEffect(() => {
-    if (
-      isUndefined(previousPersistedFilters) &&
-      !isUndefined(persistedFilters)
-    ) {
-      setFilteredInsightTypes(
-        (persistedFilters?.issueTypes as InsightType[]) ?? []
-      );
-      setFilteredInsightTypesInGlobalScope(
-        (persistedFilters?.issueTypesInGlobalScope as InsightType[]) ?? []
-      );
-      setFilteredCriticalityLevelsInGlobalScope(
-        persistedFilters?.criticalityFilterInGlobalScope ??
-          insightsInitialState.filteredCriticalityLevelsInGlobalScope
-      );
-      setFilters(persistedFilters?.filters ?? []);
-      setLastDays(persistedFilters?.lastDays ?? insightsInitialState.lastDays);
-      setAreFiltersRehydrated(true);
-    }
-  }, [
-    previousPersistedFilters,
-    persistedFilters,
-    setFilters,
-    setFilteredInsightTypes,
-    setFilteredInsightTypesInGlobalScope,
-    setFilteredCriticalityLevelsInGlobalScope,
-    setLastDays
-  ]);
-
-  // Persist filters on its change
-  useEffect(() => {
-    if (
-      (previousFilteredInsightTypes !== filteredInsightTypes ||
-        previousFilters !== filters ||
-        previousFilteredCriticalityLevels !== filteredCriticalityLevels ||
-        previousLastDays !== lastDays) &&
-      areFiltersRehydrated
-    ) {
-      setPersistedFilters({
-        ...(scopeSpanCodeObjectId
-          ? {
-              issueTypes: filteredInsightTypes ?? [],
-              issueTypesInGlobalScope:
-                persistedFilters?.issueTypesInGlobalScope ?? [],
-              criticalityFilterInGlobalScope:
-                persistedFilters?.criticalityFilterInGlobalScope ?? []
-            }
-          : {
-              issueTypes: persistedFilters?.issueTypes ?? [],
-              issueTypesInGlobalScope: filteredInsightTypes,
-              criticalityFilterInGlobalScope: filteredCriticalityLevels
-            }),
-        filters
-      });
-    }
-  }, [
-    previousFilteredCriticalityLevels,
-    filteredCriticalityLevels,
-    previousFilteredInsightTypes,
-    filteredInsightTypes,
-    previousFilters,
-    filters,
-    setPersistedFilters,
-    areFiltersRehydrated,
-    persistedFilters,
-    scopeSpanCodeObjectId,
-    lastDays,
-    previousLastDays
-  ]);
-
-  // Reset insight type, criticality and unread filters (for span and global scopes) on backend instance change
-  useEffect(() => {
-    if (
-      areFiltersRehydrated &&
-      Boolean(
-        previousBackendInfo &&
-          !areBackendInfosEqual(previousBackendInfo, backendInfo)
-      )
-    ) {
-      setFilteredInsightTypes([]);
-      setFilteredInsightTypesInGlobalScope([]);
-      setFilteredCriticalityLevels(
-        insightsInitialState.filteredCriticalityLevels
-      );
-      setFilteredCriticalityLevelsInGlobalScope(
-        insightsInitialState.filteredCriticalityLevelsInGlobalScope
-      );
-      setFilters([]);
-    }
-  }, [
-    previousBackendInfo,
-    backendInfo,
-    areFiltersRehydrated,
-    setFilteredInsightTypes,
-    setFilteredInsightTypesInGlobalScope,
-    setFilteredCriticalityLevels,
-    setFilteredCriticalityLevelsInGlobalScope,
-    setFilters
   ]);
 
   const handleJiraTicketPopupOpen = useCallback(
@@ -248,10 +95,7 @@ export const Insights = ({ insightViewType }: InsightsProps) => {
     isLoading: boolean
   ): JSX.Element => {
     const isInitialLoading =
-      (!data && isLoading) ||
-      !backendInfo ||
-      !storedInsightViewType ||
-      !areFiltersRehydrated;
+      (!data && isLoading) || !backendInfo || !storedInsightViewType;
     if (isInitialLoading) {
       return <EmptyState preset={"loading"} />;
     }
